requirements.txt
pandas>=2.0
pyyaml>=6.0
edgartools>=4.6.3
hishel<1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12
yfinance>=0.2.66
pypdf>=5.2
cryptography>=42.0
